Remove a warning from igate.c
This usage:
  if (!heard_via_tnc_in_past_hour(call)==1)

gets a warning from clang.  That's because the logical not applies
only to the function call, and not to the whole expression.  It so
happens that this is OK, but is silly.  the function in question
returns the result a boolean expression, but cast to an int, so it is
sufficient to leave off the "==1".  Everywhere else this function is
called it is called without the logical negation, and isn't compared
to  zero with "==0".  It needn't be compared to one when negated.