Fix up a warning in interface.c
isRMC and isGGA were being implicitly declared.  Unfortunately,
when *explicitly* declaring them in interface.c, we *also* get a new
warning because these two functions take a char pointer, and the thing
that's being passed to them is an "unsigned char" pointer.

To shut up the second warning, I'm just casting the unsigned char
pointer to a char*.  It'll be fine.  But with my luck, clang will
complain about that, too.