Separate script configures from each other
On at least *some* versions of autoconf, configure is throwing errors
when trying to do the script configs --- apparently it applies the
second argument of AC_CONFIG_FILES to every single step as it
processes the first argument.  That means it's trying to chmod files
that don't exist yet.

So instead of having all the scripts in one AC_CONFIG_FILES call, do
each one separately.