Expand number of builds for travis
Travis was set up to do two builds on Linux, one with lots of optional
packages, one with minimal, but both on the same OS.  The default
"linux" distro is Ubuntu Trusty, or 14.04.  That system uses gcc 4.8
as its default compiler.

Ubuntu Xenial uses a newer compiler, but not really, really new.  It
has gcc 5.4.

This commit adds a third linux build, the full build under Xenial.  No
package names have changed, so this is just a copy of the other build
with an additional "dist:" tag.