Remove blocks of commented-out code from interface.c
This file was absolutely peppered with blocks of code that had been
commented out (sometimes for more than 15 years).  This code simply
makes the file less readable, and doesn't actually help developers.
Most of the commented out code was debugging print statements, but in
a lot of cases it was stuff that had been determined to be incorrect
and replaced.

I am removing all of this code.  I'm doing this in a separate commit
so my later work in removing warnings is a cleaner commit.