Remove warnings from interface.c
interface.c threw many warnings.  Most were "variable set but unused"
warnings, which are harmless but annoying.  One was a "variable is
possibly used uninitialized".  Many were errors in the use of strncat.

strncat is designed to append one string to another, assuring null
termination of the string and allowing one to make sure that the
target buffer is never overrun:

strncat(char * restrict s, const char * restrict append, size_t count)

The last argument is supposed to be the maximum number of characters
to copy from "append" into "s" without overrunning the buffer.

However, it is extremely common in Xastir for the final argument to be
the exact number of characters in "append" instead of the actual
maximum number of bytes it is safe to copy.  This is a misuse of
strncat, and it leads to a potential for a buffer overrun (what if there
isn't ROOM for that many characters in the target?), and makes gcc
throw "stringop-overflow" warnings.

This commit fixes that sort of error in this file, by making sure to
compute the actual number of left-over bytes in the target.  In this
case, because the target is a statically dimensioned array and not
dynamically allocated, we're talking "sizeof(array)-strlen(array)-1"
as the number of unoccupied bytes in the array, and the maximum safe
transfer quantity.  It doesn't matter that the "append" string is only
one character.  What matters is if there's room to PUT one character.