Fix "toupper" misuse in interface.c
The standard C function "toupper" takes a character as argument and
returns the corresponding upper case character if there is one, or the
same character if there isn't.

It does NOT modify its argument, because the argument is passed in by
value.  From the man page, the function prototype is:
    int toupper(int c);
There is no way this could ever change its argument.

interface.c had this:

    // Change callsign to upper-case and pad out to six places with
    // space characters.
    for (i = 0; i < (int)strlen((const char *)data); i++) {
        toupper(data[i]);
    ...

As a result, toupper was in fact doing nothing at all, and data[i] was
left as lower case if it was lower case to begin with.  What should
have been done is "data[i]=toupper(data[i])".

This code has been wrong since the day it was committed, in commit
1f7c0938 on 17 July 2002.

It took building on OS X with Clang to spot this, because clang throws
a warning:

interface.c:3511:9: warning: ignoring return value of function declared with pure attribute [-Wunused-value]