Silence warnings about ignored return values
In newer Linux systems, the setegid/seteuid functions have been
declared in system headers as *requiring* that the return value be
used.  We always ignore that return value, because we don't care if
the function works or not, and don't always need it to.  This causes
GCC 5.4 and later to scream loudly in warnings.  All.  Over.  The.  Place.

In order to shut up this meaningless warning, I cahnged the
ENABLE_SETUID_PRIVILEGE and DISABLE_SETUID_PRIVILEGE macros to test
the return value of the function, then do nothing special no matter
what the test results are.  This *should* silence the warnings.