Try to silence two more interface.c warnings
We are ignoring the return value of a "write" call, and newer linux
systems hate that.  Stop ignoring it (well, keep ignoring it, but do
so in a way that makes gcc think we're not).

We are also getting a new warning on Ubuntu 16.04 (but not Ubuntu
14.04) about a variable that "might be clobbered by longjmp or vfork."
Make this variable volatile in the hopes of silencing this warning.
This technique worked elsewhere to silence such a thing.