Enable "silent rules" for make
By default, Automake creates makefiles that spew the entire
compilation and link lines during a build.  This is not usually
helpful, especially when the compile lines have lots of options on
them, and especially not at link time when there are usually tons of
options and libraries present.

AM_SILENT_RULES turns this off by default, and instead of having the
compile lines, you just get a brief version, as in:

  CC       util.o
  CC       view_message_gui.o
  CC       wx.o

and so forth.

If one is doing debugging and absolutely NEEDS to see the full line,
one can simply invoke make with an extra argument:

  make V=1

to re-enable the verbosity.

Once we get rid of all the warnings and other useless output, watching
Xastir compile should be a fairly quiet thing, with real errors and
meaningful warnings easily spotted instead of having to be dug out of
the noise.