Ensure x_spider children don't return
If the UDP x_spider server couldn't open any ports (such as due to
another xastir process having them open) the child would return to
the main control logic of xastir, leading to two copies of xastir
running and duplicate windows appearing. Add an exit(1) after the
call to UDP_Server() so this can't happen. This fixes #65.

The TCP server didn't use return, but I added an exit(1) after the
call to TCP_Server() just to be safe.

I also fixed some warnings on MacOS compiles. This helps with
issue #24.