Remove some strncat stringop-overflow warnings
GCC 8 throws warnings if you use strncat like this:

   strncat(destination,"c",1)

because it's wrong --- strncat is supposed to be used to prevent
buffer overruns in the destination, and therefore the third argument
is supposed to be the number of characters that will fit in
destination.  If destination is an array declared as:
   char destination[DESTSIZE]

then one can use "sizeof(destination)-strlen(destination)-1" for the
count, because this is how many unused characters there are in
destination, leaving room for the terminating null character.

If destination is a dynamically allocated string, then this is not
good enough, and one actually has to make sure one has the real size
of the allocated array around.  sizeof won't work there, as that will
just return the size of the pointer, not the size of the array.

At any rate, there  are lots of such warnings thrown by Xastir code.
I am fixing three of them, in objects.c and x_spider.c.

This makes these files compile without warning on my BSD machine with
gcc8.  Unfortunately, we don't have gcc8 available to us on any of the
travis CI virtual machines, so this won't show any improvement there.

I am not doing this as a pull request, because the changes are just
too trivial.