Remove three "stringop-overflow" warnings
This, as with the last commit, concerns warnings about the use of
strncat.  The third argument must be the number of free elements of
the destination array, and NOT the length of the source string.
Using strncat with the source string length in the third argument can
lead to a buffer overrrun, which gcc 8 warns about like this:

../../Xastir/src/xastir_udp_client.c: In function 'main':
../../Xastir/src/xastir_udp_client.c:198:13: warning: 'strncat' specified bound 10 equals source length [-Wstringop-overflow=]
             strncat(extra, ",-identify", 10);
             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

This is done in pursuit of issue #24.