Silence gcc warnings about unused fread return result
Linux headers now have the fread function set with a "warn unused
result" attribute, which causes gcc to spew warnings about, you
guessed it, unused return results.

This commit simply uses the return result in a conditional that then
does nothing:
   if (fread(...) != 0) {
       //
   }

I am doing this as a straight commit-and-push because it is so
trivial, and because the last PR we had that got reviewed
did *precisely* this change in several other files.

This change being made to address issue #24 in part.