Silence used result warning
GCC on Linux is throwing warnings when the return value of fgets is
ignored.  Linux system headers now define this function with a "warn
unused result" attribute.

This commit fixes these warnings in one file.  I am simply committing
and pushing without making a pull request, because the fix in this
file is precisely the same as the fix done in rac_data.c in commit
83ece25 earlier today, and which was reviewed as part of PR #68.

This commit being made to further issue report  #24 in part.