Add dependent package
The GraphicsMagick package on Ubuntu 16.04 is missing a dependency,
and so libwebp-dev doesn't automatically get installed when
libgraphicsmagick1-dev is installed --- it's a packaging blunder on
the part of the package manager.

Configure is failing to link with GM because of this missing library.
Adding this one package gets the Travis-CI Ubuntu 16.04 properly
configured, and gets it building with GM.