Reindent map_geo.c to project standards
This file threw "misleading indentation" warnings from gcc 6 and 8.

A "misleading indentation" warning is thrown for stuff like this:

  if (foo)
     bar;
     baz;
  foobie;

This is indented as if "baz" were inside the conditional, but the lack
of braces means that it really isn't.  GCC 6 and 8 flag this and warn
about it.

This commit shuts them up by re-indenting the entire file to project
standards using emacs' "indent-region" function, configured according
to project standards of four-space block indentation and no tabs.

In doing so, it corrected *many* other incorrectly indented lines (per
the project standard).

This commit does NOTHING other than reindent.