Fix "misleading indent" warnings in map_tif.c
This commit is nothing but a change in whitespace.  No code changes or
line deletions have been made.

As with my work on map_geo.c, I have addressed "misleading indent"
warnings by completely reindenting this file using Emacs'
"indent-region" function, set up to respect project style of four
space indent and no tabs (this is automatically set up by the presence
of the funky comment line at the top of the file as soon as Emacs
opens the file).

In doing so, emacs corrected a number of other indentation issues in
the file.

Unfortunately, this file has a number of commented-out blocks of
code.  Those blocks of comments inside C style comment
delimeters (/* */) were things that Emacs wanted to reindent
inappropriately (as if they were text instead of code).  I used "git
add -p" to review each change in the file, and rejected all changes
that were just changes to indentation of comments.

I would, of course, prefer that all blocks of commented out code be
REMOVED, but I have refrained from doing so so that this will be a
WHITESPACE CHANGE ONLY commit.