Make sure NO_XPM gets defined appropriately
The declaration of the "atb" variable in map_geo.c is inside an
"#ifndef NO_XPM", but the USE of this variable is inside the else
block of an "#ifdef HAVE_MAGICK/#else/#endif".  And NO_XPM was only set
if HAVE_LIBXPM and HAVE_LIBXPM_IN_XM ar both undefined.

That means that if HAVE_MAGICK is defined, but libxpm is detected by
configure, atb gets declared, but never used.

This commit simply makes NO_XPM be defined to 1 if HAVE_MAGICK is
defined OR if we can't find XPM.

It removes one more warning about unused variables, still in hot
pursuit of bug #24.  I have only seen this warning under GCC 8 on BSD
and GCC 6 on Raspbian.  It does not show up on any of our Travis
builds.