Make sure XPM calls only happen if XPM configured
The last commit made sure that none of the variables needed by the XPM
code are allocated if XPM won't be used, but a long-standing mistake
in ifdefs did not assure that the XPM functions are not called if XPM
didn't get found by configure.  This would have resulted in link
failures of Xastir at the end of the build, when XpmReadFileToImage
never got found.

This commit wraps all of the Xpm code (which itself is in the #else
block of an #ifdef HAVE_MAGICK conditional) is wrapped by the same
ifdef that wraps the declaration of the variable it uses.  If neither
HAVE_MAGICK nor the various macros required to locate XPM are set, we
instead emit an error message about not being able to display the
chosen map file.