Fix deprecated GetIndexes usage
In commits b071954 and 7815dd9 (and possibly others associated with
PR #40), the GetIndexes function (present in both ImageMagick and
GraphicsMagick, and deprecated in both) was coded around so that if
we've got newer versions of the library, we use the new API, and if we
have an older version, we use the older API.

But #40 only fixed these deprecated uses in a few files, and mised the
one usage in map_geo.c

This commit copies over the fix verbatim from map_OSM.c, and shuts up
the deprecation warning.

Fixes #81.