Fix a misleading indent warning in main.c
There was one misleading indentation warning around line 6490, where
lines following an un-braced if statement were indented as if they
were part of the if block.  GCC 8 warns about this.

I fixed the two lines that were incorrectly indented and reformatted
the function arguments to match the lines near them.

I tried to run this entire file through emacs' indent-region function,
and wound up with a huge mess (something is confusing it about the
syntax).  Rather than make a sweeping commit that changes whatever it
is that is confusing the indenter, I just fixed the two problematic
lines to silence the warning.

This commit changes nothing but a handful of lines of white space, so
I'm not bothering with a PR and code review.  This commit helps
address bug report #24.