Expand "xfontsel" char string
xfontsel is declared as a 50 character array in Map_font_xfontsel, but
the sprintf in the same function (which doesn't guard against buffer
overruns) can write 51-61 characters in there (including the null
string), according to GCC8 on BSD, where XFONTSEL_PATH is
"/usr/local/bin/xfontsel":

../../Xastir/src/main.c: In function 'Map_font_xfontsel':
../../Xastir/src/main.c:4560:42: warning: 'sprintf' writing a terminating nul past the end of the destination [-Wformat-overflow=]
         "%s -print -title \"xfontsel %d\"",
                                          ^
../../Xastir/src/main.c:4559:5: note: 'sprintf' output between 51 and 61 bytes into a destination of size 50
     sprintf(xfontsel,
     ^~~~~~~~~~~~~~~~~
         "%s -print -title \"xfontsel %d\"",
         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
         XFONTSEL_PATH,
         ~~~~~~~~~~~~~~
         fontsize);
         ~~~~~~~~~
23 chars for XFONTSEL path, 26 fixed characters for the format, and
several more for the integer font size can easily exceed 50 chars.

I am expanding the array to 70 chars to prevent that buffer overrun
warning (and the possibility of a buffer overrun) on this system.  We
should probably bump it even further to allow for the fact that
XFONTSEL_PATH could be just about anything, and not presume that it
will be a short path.

Not bothering with a PR for this one character change.

Silences warnings on at least one system, in pursuit of bug #24.