Remove "tautological compare" warning
A fix in PR #87 removed a "set but unused" warning from GCC, but
introduced a "tautological compare" (i.e. always true by definition)
warning from Clang.

Since the actual code doesn't care whether it's true or not, make one
that's always going to be false, but not obviously so to the
compiler (lest we get a "never can be true" warning, which Clang likes
to do, too).