Remove test of array name against NULL
When an array is declared as "char array[SIZE]", array is *technically* a C
pointer, but unlike pointers  declared as "char *" they can never be
null.

Clang throws warnings whenever it encounters something like:
  if (array != NULL)

because the condition can never be false.

I am removing one such conditional in main.c, because
predefined_object_definition_filename is declared in objects.h as a
character array, and so it is not a dynamically allocated string whose
pointer must be checked before access.

The commit removes one conditional line, and one line with a bracket
on it, then reindents what used to be inside the conditional.