Remove compare of array name to null
This commit removes another instance a warning from Clang about
comparing a statically sized array to NULL as if it were a dynamic
pointer.

In this case, we were doing so in the conditional part of a ternary
operator.  It's the "tautological pointer compare" that is getting
whined about:

map_cache.c:833:14: warning: comparison of array 'mc_delete_file' equal to a null pointer is always false [-Wtautological-pointer-compare]
            (mc_delete_file == NULL) ? "(null)" : mc_delete_file,

Committed directly because it is so trival and should not require a
PR.  Knocking off one more warning for bug #24.