Reindent dlm.c to project standards
Jason may be on the track of a thread-unsafe issue in this code, and
I went over to dlm.c to look at it.  I found a number of
lines (including the very one I was curious about) were not indented
right.

I ran the whole thing through Emacs' auto-indenter.  I rejected its
attempt to reformat a large comment block, but otherwise it made quite
a few corrections to the indentation.