Fix use after free bug in Station_List_fill
Station_List_fill gets called repeatedly. Each run it
would add a callback to the icon button. On the next run
it would free the string that was passed to the callback
but not delete the old callback. Over time we would develop
a stack of old callbacks on the icon button that would
point to freed strings.

Add logic to remove the existing callback before we free
the string from the previous run.

Fixes #88