Revert "Fix buffer overflow in jump_sort()"
This reverts commit 57735df7783f761479d4928b152fa86ae989f008.

This commit, while it looked correct at cursory glance, actually
causes an inconsistency between how locations_db.dat is written and
how it is read.

I am not fixing the potential buffer overrun that the commit was
intended to fix just yet, but I'm reverting this commit because it
breaks reading of the map display bookmarks files.