Make AC_DEFINES define constants to 1
In many places in Xastir's configure.ac, we were invoking the macro

  AC_DEFINE(<symbol>, , Define to 1 if have <thing controlled by symbol)

There is nothing strictly wrong with this --- it causes configure to
add the following thing into config.h.in:

  /* Define to 1 if have <thing controlled by symbol> */
  /* #undef <symbol> */

And if <thing controlled by symbol> is found, causes config.h to have
this:

  /* Define to 1 if have <thing controlled by symbol> */
  #define <symbol>  /**/

That is, it defines the symbol, but gives it no value.

This has worked fine in Xastir forever, because we only use <symbol>
in ifdefs, and never check its value.  So defining it is enough.

But the comments, which are just copied verbatim from the third
argument of AC_DEFINE, are misleading, and led to confusion today when
someone misread "#define <symbol> /**/" for "Configure isn't finding
the feature controlled by <symbol>, and it should have."

This commit makes all those places were we're saying "Define to 1"
actually cause the symbol to be defined as 1.  This has NO effect on
program correctness, but making the comments less misleading should be
a good thing.

A completely equivalent approach would have been to leave it as it
was, but change the COMMENT from "Define to 1 if" to "Define if."  I
decided this approach could still lead to confusion (due to the "/**/"
present in config.h), so chose to change the definition, not the
comment about the definition.