Reorganize probe for Xpm, output summary
Curt experienced a problem today with configure not recognizing Xpm
libraries, and didn't notice it until PNG snapshots didn't work.  The
"not recognizing" part turned out to be a PEBKAC issue, but the fact
that there was basically silent failure to detect XPM and no warnings
isn't good.

To address this, I've added information about the results of the Xpm
probe down in the main summary area, just like we do for Magick and
other recommended options.

I have also rearranged how we probe for Xpm.  There was a fairly
confused probe for both X11/Xpm.h and Xm/XpmI.h, and associated
functions XpmWriteFileFromPixmap and XmeXpmWriteFileFromPixmap.  It is
clear this was intended to handle an odd library arrangement where Xpm
itself didn't exist, but libXm contained a compatible function.

I grepped the code, and we NEVER call XmeXpmWriteFileFromPixmap, so it
is unlikely this still works.  Xm/XpmI.h does not appear to exist in
any modern build of Motif, and appears to have been replaced with
XpmP.h, which #defines XpmWriteFileFromPixmap to point at the Xme
version.

Now, we probe for X11/Xpm.h, and if it's found, we try to find
XpmWriteFileFromPixmap in -lXpm.  If that works, we declare victory
and don't keep trying.  Only if X11/Xpm.h doesn't exist do we then try
to look for Xm/XpmI.h and XmeWriteFileFromPixmap.  If neither works,
we emit a warning just like we do for missing shapelib.

If either works, we emit a note in the summary that Xpm/Snapshots will
be built in.

We should probably revisit whether the probe for Xm/XpmI.h and the
XmeXpmWriteFileFromPixmap function should remain, as the header does not
appear to exist in any OpenMotif install I've found, and this is
probably never getting exercised in any builds.