Replace per-file emacs variable setting with directory vars
Some files in Xastir had (at my hand) been setting per-file emacs
variables in order to enforce project indentation standards that
overrode my own defaults.

This commit removes all of them and replaces them with a
directory-local setting of the same variables.

In this way, the files themselves are left alone, and the standards
are applied to all files for emacs users by the presence of the
.dir-locals.el file.