Replace tabs with spaces
This commit does nothing but replace tabs with spaces, and in a small
handful of cases, some lines that had no whitespace at the beginning
to line up with other lines in the same block.

This commit is so utterly trivial I am NOT doing a pull request for
it.

This is in support of issue #112