Fix indentation and brace style in datum.c
This file needed some work on reindentation.  It also did not conform
to the projects brace style (K&R style, or "egyptian" style, or, as I
prefer to say "icky style") in a few places.

There was a large block of initializers at the top that I did NOT
reindent, as it would have caused most of the content to be so far to
the right that it would all be wrapped to the next line.  I just left
those lines alone.

Other than the handful of braces moved, this is a whitespace only
commit.

Done in support of issue #112.