Updating lang files for %d variable
We were passing an integer to the language function but the strings
defined were not capable of using it, which most likely would cause
a segfault were they actually to be used. While at it I corrected
a string in the Dutch file that was incorrect: Was a copy of string
023 in that section instead of the INVALID ENUM string it should
have been.