Update emacs directory-local vars to new coding standard
We are no longer going to be using 4-char indent going forward in the
project.  While we're not reindenting everything before the next
release, Curt already changed the vim config in PR #115, so here's the
change for the emacs config.

I am not doing this via a PR, because that's pointless.  This is a
simple change, and I'm just pushing it.

Consistent with #115.