Change code to quiet compiler warnings
This is in support of Issue #24 (already closed). These changes
quiet compiler warnings on gcc7 and gcc8. Most of the changes have
to do with xastir_snprintf() calls and strings having the chance to
overrun the destination. One change to the Travis-CI config takes
out the compiler flags that quieted the warnings so we can see when
further issues appear from this date forward.