Make Xastir recognize GGA and RMC sentences from any talker
NMEA sentences are:
  $<talker><sentencetype>,...

where <talker> is a two-character identifier representing the device
type.  The rest of the string is defined in the standard and doesn't
depend on the talker id.

Until commit 2264a035 (which merged PR#58), Xastir only recognized the
GP talker (GPS).  PR #58 added support for GN, GA, GL and GB.  But
there are many other talker IDs out there, and issue #121 calls for
yet another, $II.

This commit simply makes Xastir ignore the talker ID and look only for
the pattern "$..GGA" or "$..RMC" (where "." is regexp-speak for "any
character").

Issue #121
Issue #53
PR #58