Make "rtree" the only shapefile reading method
Years ago, Xastir's shapefile support always read an entire shapefile
if any part of the shapefile's bounding box intersected the current
map screen.  That could be enormously expensive when a shapefile had a
very large number of objects in it, only a small number of which were
actually on screen.

In 2004 (commit 7c9a474) I added an experimental feature where instead
of sequentially reading an entire shapefile only to find that only a
few objects in it were visible, a spatial index (R-tree data
structure) would be built from a shapefile the first time it was
read.  From that point forward, all reading of the file is done by
searching the spatial index for shapes that intersect the screen's
bounding box, and these individual objects are read via random access.
At that time, because it was an experimental feature, it was made a
configure option (--with-rtree) until its value could be proven.

In 2006 (commit cf417d1), the value had been proven well enough that the default
behavior of configure was to enable rtree, and one had to *disable*
it (via "--without-rtree") if one didn't want it.  Disabling it just
slowed down the shapefile rendering (often by a lot) so it is pretty
much not done anymore.  I've spotted a number of packages out there
that explicitly include "--with-rtree" even though this is already
the default, presumably to make absolutely sure this optimization was
included.

After a little discussion, and 13 years of experience showing that the
rtree code is robust and an improvement over the older method, we're
making rtree the ONLY code for handling shapefiles.  The ifdef
handling the rtree stuff is now gone, as is the code that was selected
when "USE_RTREE" was not defined, as are the conditional compilation
options.

This commit closes issue #127.