Remove "--without-rtree" from travis config
My removal of the non-rtree code requires that we stop trying to build
with the "--without-rtree" option in the "minimal build" test on
travis.  rtree is now unconditional on this branch.

(Note:  I had assumed that a travis-ci failure I saw on the "exercise
configure options" build was due to this, but it was in fact just a
github clone hiccup)