Fix compilation failure with newer GraphicsMagick
David Flood reported today that Xastir failed to compile for him,
because of a missing file "magick/image-private.h", which was getting
included by a GraphicsMagick header.  Turns out this same issue has
been cropping up in distro-specific bug reports all over the place
lately:

https://bugzilla.redhat.com/show_bug.cgi?id=1722840
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=924329#18

And today, I tried to compile Xastir on FreeBSD only to find that on
June 20th my own system had updated GraphicsMagick that exposed the
same issue.

This turns out to be a combination of mistakes in Xastir that have
come back to bite us.

In commit  4210671c37 back in 2007, an attempt to assure that map_geo
consistently released memory whenever it exited introduced a call to
the private function DestroyImagePixels, which was meant only to be
called by the DestroyImage function.  At the same time, great care was
taken to make sure that DestroyImage was called at all exit points, so
the call to DestroyImagePixels was unnecessary.

Later, GM removed the public declaration of DestroyImagePixels and hid
it behind an ifdef that was intended to be set only by GraphicsMagick
source code.  This led to a compilation warning in Xastir because the
function, which is supposed to be "void DestroyImagePixels(Image
image)" was used but implicitly declared as an int function with no
prototype.  To address this, instead of removing the unnecessary call,
a definition of the private "MAGICK_IMPLEMENTATION" macro was added to
assure that DestroyImagePixels' declaration was picked up.  That was
done in 2010 in commit f350db4.

Since "draw_geo_image_map" is the only place where this function was
ever called, and because all exit points of the function correctly
call DestroyImage (which itself internally calls DestroyImagePixels),
I am removing the inappropriate use of the private function *and* the
definition of the MAGICK_IMPLEMENTATION macro.  This should have no
impact on memory bloat (since all returns of the function call
DestroyImage), and should end the compilation problem for all distros
for good.