Make Travis use later Xcode on OS X
Starting in June, Travis-CI started failing to run homebrew on OS X,
saying "Your homebrew is outdated..."  This is not unique to Xastir:
https://travis-ci.community/t/xcode-8-3-homebrew-outdated-error/3798

When I pushed code changes last night, the OS X build failed, and I
keep getting annoying emails about it still failing.  I am trying to
make travis use a later image with an updated xcode instead, in the
hopes that this will stop.

If it does not stop, I may have to add the "update: true" tag to the
homebrew section, but this can *seriously* slow down the build, so I
am trying to avoid it.