Remove some ImageMagick 6 deprecation warnings
map_geo.c and main.c were still using deprecated functions in
ImageMagick 6.  These functions have actually been REMOVED in
ImageMagick 7, so we really should not be using them anymore.

Configure is currently set up to refuse to use any ImageMagick in the
7 series because when I tried to build Xastir with it it wouldn't
work, but that may be only because my ImageMagick 7 was built with
HDRI support (we have NEVER supported that feature in Xastir, and
probably never will).  HDRI support completely changes datatypes of
things to be incompatible with how Xastir is written.

Addresses #150