Remove non-dbfawk shapefile rendering
Before dbfawk was written by Alan Crosswell and added to Xastir in
commit 7ac9b21 (and subsequent commits) in July 2003, Xastir was only able
to render shapefiles that had been specifically hard-coded into
map_shp.c.  Any shapefile that was not one of those hard-coded files
got rendered as plain black lines.  It was not easy to add hard-coded
rendering to the file, and each time a new block was added, the file
got substantially larger.

DBFAWK off-loads the rendering rules to an external, "awk" like
control file that is read at run time when the shapefile is drawn.

It has now been 16 years since Alan first added that code, and almost
all of our users make use of it.  Few, if any, use the old code, which
was only built if one explicitly disabled dbfawk OR did not have the
PCRE library.

This commit makes PCRE a mandatory third party library IF shapefile
support is requested, and removes the capability for disabling dbfawk.

Configure now probes this way:

   if (user has requested shapefile support)
     {
        check for shapelib
        flag whether we found it
     }
   if (user has requested shapefile support, and we found shapelib)
   {
     check for pcre
     error out if we have not found it, telling user to rerun
       configure with "--without-shapelib" or to install PCRE
   }
   if (user has requested shapefile support, and we haven't found
     shapelib)
   {
      Notify user that we're disabling shapefile support
   }

All vestiges of the code that would be compiled if WITH_DBFAWK were
not defined has been removed, and all ifdefs corrected to assume
DBFAWK can be used of both LIBSHP and LIBPCRE ifdefs are set.

I have tested this the following ways:
  1) normal, full build that I always do
  2) build with "--without-shapelib", even with libshp installed
  3) build with shapelib installed and requested, but pcre uninstalled

All three behave as I expect:
  1) runs fine with my normal DBFAWK-enabled build
  2) turns off shapelib support, doesn't even look for pcre
  3) finds shapelib, but configure aborts with informative message.

Closes #128

https://github.com/Xastir/Xastir/issues/128