Remove borken tests from travis
I have removed the "--without-dbfawk" and "--without-pcre" options
from configure, and now only "--with/without-shapelib" exists.  If one
has shapelib and not pcre, it's an error, and if one asks for no
shapelib, we don't bother probing for pcre.  dbfawk is always on if
shapelib is.

Done as part of bug #128