Update aircraft bit-strings + make more efficient
It's been a while since the bit-strings were checked against the
current assignments and the Mode-S group "Countries.dat" file.
Also the way the coding was done originally it was one big if/else...
Now it is more efficient as it checks shorter strings first by
country, then if a country match, goes into a block which has
country-specific additional strings to check.