Remove "Xastir_tigerpoly.py" script
Xastir_tigerpoly.py was a script that used GDAL python bindings to
convert the ancient topological format TIGER/LINE map files to simple
shape shapefiles.

Apparently this script requires python 2 (which is no surprise,
because it was written years before there was a python 3), and at
least one distro has made their Xastir package depend on python2
solely because of this script.  As a result, Xastir is now slated for
removal from that distro's packages because they're removing python2.
A simpler approach would just be to drop the python dependency and let
this script simply not work, but that is not what was being done.

I am therefore removing this script from the Xastir distribution.  The
old TIGER/LINE format was discontinued years ago, and is of little
interest to anyone.  And anyone *that* interested in ancient mapping
data can resurrect this script from git history.