Temporarily disable Davis/Lacrosse testing on OS X
The latest MySQL package on homebrew (8.0.18) has a problem where it
conflicts with libprotobuf (which is apparently also indirectly required by
Xastir and is installed as a dependency).  That causes our CI builds
to report failure when trying to build LaCrosse and Davis daemons.

The 8.0.18 issue is a known problem and is supposedly fixed in 8.0.19.
So we're temporarily disabling build testing of the Davis and Lacrosse
daemons on OS X until 8.0.19 is released and available under homebrew.